Frontend/requirements.txt
Flask
python-dotenv
requests
flask-login
PyJWT
json

